fix2406302150

+++ src/users/useCase/banUnbanUser-use-case.ts
@@ -37,13 +37,15 @@ export class BanUnbanUserUseCase implements ICommandHandler<BanUnbanUserCommand>
 			])
 		}
 		const banUser = await this.usersRepository.banUser(command.id, command.banInputInfo)
-		
-		await this.deviceRepository.deleteAllSessions(command.id);
-		await this.blogsRepository.banBlogByUserId(command.id, command.banInputInfo.isBanned)
-		await this.postsRepository.banPostByUserId(command.id, command.banInputInfo.isBanned)
-		await this.commentRepository.banComments(command.id, command.banInputInfo.isBanned);
-		await this.likesRepository.banCommentLikes(command.id, command.banInputInfo.isBanned);
-		await this.postsRepository.banPostLikes(command.id, command.banInputInfo.isBanned);
+		if(banUser.isBanned) {
+			await this.deviceRepository.deleteAllSessions(command.id);
+			await this.blogsRepository.banBlogByUserId(command.id, command.banInputInfo.isBanned)
+			await this.postsRepository.banPostByUserId(command.id, command.banInputInfo.isBanned)
+			await this.commentRepository.banComments(command.id, command.banInputInfo.isBanned);
+			await this.likesRepository.banCommentLikes(command.id, command.banInputInfo.isBanned);
+			await this.postsRepository.banPostLikes(command.id, command.banInputInfo.isBanned);
+			return 
+		}
 		return 
 	}
 }
